requirements.txt
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
Flask-Session==0.6.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pillow==10.2.0
qrcode-artistic==3.0.2
requests==2.31.0
segno==1.6.0
urllib3==2.1.0
Werkzeug==3.0.1
